chore: add ResponseEntity to requests

+++ src/main/java/com/codecool/budapestgo/service/StopService.java
@@ -35,16 +35,28 @@ public class StopService {
                 .orElseGet(() -> ResponseEntity.notFound().build());
     }
 
-    public void deleteStopById(Integer id) {
-        stopRepository.deleteById(id);
+    public ResponseEntity<String> deleteStopById(Integer id) {
+        Optional<Stop> searchedStop = stopRepository.findById(id);
+        if(searchedStop.isPresent()) {
+            stopRepository.deleteById(id);
+            return ResponseEntity.ok("Stop deleted.");
+        }
+        return ResponseEntity.status(HttpStatus.NOT_FOUND)
+                .body("Stop with" + id + " don't exist.");
     }
 
-    public void addStop(StopDTO stopDTO) {
-        Stop stop = Stop.builder()
-                .name(stopDTO.name())
-                .location(new Point(stopDTO.latitude(),stopDTO.longitude()))
-                .build();
-        stopRepository.save(stop);
+    public ResponseEntity<String> addStop(StopDTO stopDTO) {
+        Optional<Stop> searchedStop = stopRepository.getStopByName(stopDTO.name());
+        if(searchedStop.isEmpty()) {
+            Stop stop = Stop.builder()
+                    .name(stopDTO.name())
+                    .location(new Point(stopDTO.latitude(), stopDTO.longitude()))
+                    .build();
+            stopRepository.save(stop);
+            return ResponseEntity.ok("Stop created");
+        }
+        return ResponseEntity.badRequest()
+                .body("Stop already exist.");
     }
 
     public ResponseEntity<String> updateStop(StopDTO stopDTO) {
@@ -54,6 +66,7 @@ public class StopService {
             stopRepository.save(stop.get());
             return ResponseEntity.ok("Stop updated");
         }
-        return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Stop not found");
+        return ResponseEntity.status(HttpStatus.NOT_FOUND)
+                .body("Stop not found");
     }
 }
